Add an integration test for `stripUrlParams`
This test provides confidence that the method is effective in context.

In doing so it made sense to add the ability to configure the window.location.href property in Jest. Usually this isn't possible because JSDom doesn't support navigation, but the updated environment allows it on a case-by-case basis

Additionally I've added several TS packages for better documentation of DoubleClick features and more support for Jest and JSDom